requirements.txt
cpe==1.2.1
python-dateutil==2.7.2
redis==2.10.6
six==1.11.0
